requirements.txt
python-decouple
ipdb
pytz